requirements.txt
aiohttp>=3.9
flask>=3.0
python-dotenv>=1.0
py-clob-client>=0.17